+++ CHANGELOG.md
@@ -6,8 +6,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-MD5 ETag comparison for Put Object node
-MD5 ETag comparison for Put Objects node
+- Refactored List Bucket node
+- Refactored List Objects node
+- Refactored Get Object node
+- Refactored Create Bucket node
+- Refactored Put Object node
+- Refactored Put Objects node
+- Refactored Delete Object node
 ### Added
 
 - Changelog